# Conflicts:
#	README.md

+++ Phoenix/Core/ECS/include/EntityComponent.h
@@ -1,7 +1,8 @@
  #pragma once
  namespace Phoenix
  {
-  using ComponentId = size_t;
+  using ComponentId = long long int;
   using ComponentSystemId = size_t;
-  using EntityId = size_t;
+  // EntityId is a long long int because -1 is used to represent an invalid entity
+  using EntityId = long long int;
  }
